:adhesive_bandage: Update remoterun to use eiffel's shell instead.

+++ bin/remoterun-future.sh
@@ -61,7 +61,6 @@ main() {
         esac
     done
 
-
     # Check that the experiment directory exists.
     if [ ! -d "$EXPERIMENT" ]; then
         echo "Experiment directory '$EXPERIMENT' does not exist."
@@ -98,51 +97,68 @@ main() {
     COMMAND="eiffel --config-dir $EXPERIMENT $NOTIFYARGS"
     
     REPO=$(git rev-parse --show-toplevel) || { echo "Could not find the root of the git repository. Aborting."; exit 1; } 
+    cd "$REPO" || { echo "Could not change to the root of the git repository. Aborting."; exit 1; }
 
+    # Run the experiment
     if [[ -n "$DEBUG" ]]; then
         set -x
 
         # push the current changes to git
         echo "Pushing current changes to the $PUBLISHBRANCH branch..."
         git add .
-        git commit --allow-empty -am ":rocket: Auto commit before running experiment '$EXPERIMENT' on '$TARGET'." || { echo "Could not commit changes. Aborting."; exit 1; }
-        COMMIT=$(git rev-parse HEAD)
-        git push --force origin "$COMMIT:$PUBLISHBRANCH" || { echo "Could not push changes. Aborting."; exit 1; }
-        git reset --soft HEAD~1 || { echo "Could not reset changes. Aborting."; exit 1; }
+        if [[ -n $(git status --porcelain) ]]; then
+            git commit -am ":rocket: Auto commit before running experiment '$EXPERIMENT' on '$TARGET'." || { echo "Could not commit changes. Aborting."; exit 1; }
+            COMMIT=$(git rev-parse HEAD)
+            git push --force origin "$COMMIT:$PUBLISHBRANCH" || { echo "Could not push changes. Aborting."; exit 1; }
+            git reset --soft HEAD~1 || { echo "Could not reset changes. Aborting."; exit 1; }
+        else
+            COMMIT=$(git rev-parse HEAD)
+        fi
 
+        RUNDIR="/tmp/eiffel/$COMMIT"
+        ssh "$TARGET" mkdir -p "$RUNDIR/$EXPERIMENT" || { echo "Could not create the run directory on the remote host. Aborting."; exit 1; }
         # copy the experiment to the remote host
-        echo "Copying experiment '$EXPERIMENT' to '$TARGET'..."
-        rsync -azC --progress --exclude='.git' --exclude="*.pkl" --partial --inplace "$REPO/" "$TARGET:~/Workspace/phdcybersec/fl-ids" || { echo "Could not copy the experiment to the remote host. Aborting."; exit 1; }
+        echo "Copying experiment '$EXPERIMENT' to '$TARGET:$RUNDIR/$EXPERIMENT'..."
+        rsync -azC --progress --exclude='.git' --exclude="*.pkl" --partial --inplace "$EXPERIMENT/" "$TARGET:$RUNDIR/$EXPERIMENT" || { echo "Could not copy the experiment to the remote host. Aborting."; exit 1; }
 
         # run the experiment on the remote host
         echo "Running experiment '$EXPERIMENT' on '$TARGET'..."
+        echo "Running: $COMMAND"
         ssh "$TARGET" bash << EOF || { echo "Issue on remote host."; exit 1; } && echo "Experiment '$EXPERIMENT' running on '$TARGET', no apparent errors."
-cd ~/Workspace/phdcybersec/fl-ids
+set -x
+cd $RUNDIR
 tmux new-session -d -s "auto-remoterun" "nix develop $FLAKE --command $COMMAND || read && exit" || { echo "Tmux session already running."; echo "Check it out with 'tmux attach -t auto-remoterun'."; exit 1; } 
 EOF
+
     else
         # push the current changes to git
         echo "Pushing current changes to the $PUBLISHBRANCH branch..."
         git add .
-        git commit --allow-empty -am ":rocket: Auto commit before running experiment '$EXPERIMENT' on '$TARGET'."  >/dev/null 2>&1 || { echo "Could not commit changes. Aborting."; exit 1; }
+        if [[ -n $(git status --porcelain) ]]; then
+            git commit -am ":rocket: Auto commit before running experiment '$EXPERIMENT' on '$TARGET'."  >/dev/null 2>&1 || { echo "Could not commit changes. Aborting."; exit 1; }
+            COMMIT=$(git rev-parse HEAD)
+            git push --force origin "$COMMIT:$PUBLISHBRANCH" >/dev/null 2>&1 || { echo "Could not push changes. Aborting."; exit 1; }
+            git reset --soft HEAD~1 >/dev/null 2>&1 || { echo "Could not reset changes. Aborting."; exit 1; }
+        fi
+
         COMMIT=$(git rev-parse HEAD)
-        git push --force origin "$COMMIT:$PUBLISHBRANCH" >/dev/null 2>&1 || { echo "Could not push changes. Aborting."; exit 1; }
-        git reset --soft HEAD~1 >/dev/null 2>&1 || { echo "Could not reset changes. Aborting."; exit 1; }
-        
-        # switch to a new branch, create it if it doesn't exist
-        echo "Copying experiment '$EXPERIMENT' to '$TARGET'..."
-        rsync -azC --exclude='.git' --exclude="*.pkl" --partial --inplace "$REPO/" "$TARGET:~/Workspace/phdcybersec/fl-ids" >/dev/null 2>&1 || { echo "Could not copy the experiment to the remote host. Aborting."; exit 1; }
+        RUNDIR="/tmp/eiffel/$COMMIT"
+        ssh "$TARGET" mkdir -p "$RUNDIR/$EXPERIMENT" >/dev/null 2>&1 || { echo "Could not create the run directory on the remote host. Aborting."; exit 1; }
+        # copy the experiment to the remote host
+        echo "Copying experiment '$EXPERIMENT' to '$TARGET:$RUNDIR/$EXPERIMENT'..."
+        rsync -azC --progress --exclude='.git' --exclude="*.pkl" --partial --inplace "$EXPERIMENT/" "$TARGET:$RUNDIR/$EXPERIMENT" >/dev/null 2>&1 || { echo "Could not copy the experiment to the remote host. Aborting."; exit 1; }
 
         # run the experiment on the remote host
+
         echo "Running experiment '$EXPERIMENT' on '$TARGET'..."
-        ssh "$TARGET" bash << EOF || { echo "Issue on remote host."; exit 1; } && echo "Experiment '$EXPERIMENT' running on '$TARGET', no apparent errors."
-cd ~/Workspace/phdcybersec/fl-ids
-tmux new-session -d -s "auto-remoterun" "nix develop $FLAKE --command $COMMAND || read && exit" || { echo "Tmux session already running."; echo "Check it out with 'tmux attach -t auto-remoterun'."; exit 1; } 
+        ssh "$TARGET" bash << EOF >/dev/null 2>&1 || { echo "Issue on remote host."; exit 1; } && echo "Experiment '$EXPERIMENT' running on '$TARGET', no apparent errors."
+cd $RUNDIR/$EXPERIMENT
+tmux new-session -d -s "auto-remoterun" "nix develop $FLAKE --command $COMMAND || read && exit" || { echo "Tmux session already running."; echo "Check it out with 'tmux attach -t auto-remoterun'."; exit 1; }
 EOF
     fi
 
 }
 
 main "$@"
-
+cd - >/dev/null 2>&1 || exit 1
